+++ lib/src/graphql/gql_generator/annotations.dart
