+++ lib/src/graphql/gql_generator/builders.dart
@@ -1,10 +1,10 @@
 import 'package:build/build.dart';
-import 'package:dart_cast/src/gql_generator/registries/input_registry.dart';
-import 'package:dart_cast/src/gql_generator/registries/mutation_registry.dart';
-import 'package:dart_cast/src/gql_generator/registries/query_registry.dart';
-import 'package:dart_cast/src/gql_generator/registries/type_registry.dart';
 import 'package:source_gen/source_gen.dart';
 
+import './registries/input_registry.dart';
+import './registries/mutation_registry.dart';
+import './registries/query_registry.dart';
+import './registries/type_registry.dart';
 import 'generators.dart';
 
 Builder graphQlSchemaBuilder(BuilderOptions options) {
