+++ lib/src/graphql/gql_generator/exceptions.dart
