+++ lib/src/graphql/gql_generator/registries/input_registry.dart
