+++ lib/src/graphql/gql_generator/registries/registry_helpers.dart
