+++ lib/src/graphql/gql_generator/registries/type_registry.dart
