+++ test/gql_generator/mutation_registry_test.dart
@@ -1,7 +1,7 @@
-import 'package:dart_cast/src/gql_generator/exceptions.dart';
-import 'package:dart_cast/src/gql_generator/registries/input_registry.dart';
-import 'package:dart_cast/src/gql_generator/registries/mutation_registry.dart';
-import 'package:dart_cast/src/gql_generator/registries/type_registry.dart';
+import 'package:dart_cast/src/graphql/gql_generator/exceptions.dart';
+import 'package:dart_cast/src/graphql/gql_generator/registries/input_registry.dart';
+import 'package:dart_cast/src/graphql/gql_generator/registries/mutation_registry.dart';
+import 'package:dart_cast/src/graphql/gql_generator/registries/type_registry.dart';
 import 'package:gql/ast.dart' as ast;
 import 'package:test/test.dart';
