+++ test/gql_generator/type_registry_test.dart
@@ -1,6 +1,6 @@
 import 'package:analyzer/dart/element/element.dart';
-import 'package:dart_cast/src/gql_generator/exceptions.dart';
-import 'package:dart_cast/src/gql_generator/registries/type_registry.dart';
+import 'package:dart_cast/src/graphql/gql_generator/exceptions.dart';
+import 'package:dart_cast/src/graphql/gql_generator/registries/type_registry.dart';
 import 'package:gql/ast.dart' as ast;
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
